requirements.txt
openai
python-dotenv
Flask
markdown
azure-ai-inference
langchain
mysql-connector-python
langchain_community
langchain_openai
pyodbc

